refactor: Align Face: new vertex format
instead of iterating over bm.verts and storing the result in a list
it is now stored as a list of tuples, the tuple containing both the index and the result.

This way there is no reliance on the position in the list to relate it to the index of a vertex.